Create README.md
This is just a demo project for the students/newbies who are learning PHP or can be equally helpful for the college students who are in their early semester to build a web-based project where Sign up & Login pages are required simultaneously.  We here create the same pages (using HTML and CSS) in order to assist you on your project work or would be easier to understand in an easy way how it works so that other students having similar interests or learner at beginner phase can learn to code.